@@ -1,19 +1,21 @@
 package com.bassamworks.tunify.activities
 
 import android.content.Intent
+import android.content.pm.PackageManager
 import android.os.Bundle
-import android.support.design.widget.NavigationView
-import android.support.v4.app.Fragment
-import android.support.v4.view.GravityCompat
-import android.support.v7.app.ActionBarDrawerToggle
-import android.support.v7.app.AppCompatActivity
 import android.view.Menu
 import android.view.MenuItem
+import androidx.appcompat.app.ActionBarDrawerToggle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.view.GravityCompat
 import com.bassamworks.tunify.R
+import com.bassamworks.tunify.extentions.hasPermission
 import com.bassamworks.tunify.fragments.FoldersFragment
 import com.bassamworks.tunify.fragments.HomeFragment
 import com.bassamworks.tunify.fragments.LibraryFragment
 import com.bassamworks.tunify.fragments.PlayQueueFragment
+import com.google.android.material.navigation.NavigationView
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.activity_main_content.*
 
@@ -33,7 +35,10 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         nav_view.setNavigationItemSelectedListener(this)
 
+        checkPermissions()
+
         savedInstanceState ?: replaceFragment(HomeFragment())
+
     }
 
     override fun onBackPressed() {
@@ -84,9 +89,31 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         return true
     }
 
-    private fun replaceFragment(fragment: Fragment) {
+    private fun replaceFragment(fragment: androidx.fragment.app.Fragment) {
         supportFragmentManager.beginTransaction()
                 .replace(R.id.main_fragment_container, fragment)
                 .commit()
+    }
+
+    private fun checkPermissions() {
+        if (!hasPermission(android.Manifest.permission.READ_EXTERNAL_STORAGE)) {
+            ActivityCompat.requestPermissions(this,
+                    arrayOf(android.Manifest.permission.READ_EXTERNAL_STORAGE), RC_READ_EXTERNAL_STORAGE)
+        }
+    }
+
+    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>, grantResults: IntArray) {
+        when (requestCode) {
+
+            RC_READ_EXTERNAL_STORAGE -> {
+                if (grantResults[0] != PackageManager.PERMISSION_GRANTED) finish()//TODO:Show a message that this is required to continue
+            }
+
+            else -> super.onRequestPermissionsResult(requestCode, permissions, grantResults)
+        }
+    }
+
+    companion object {
+        private const val RC_READ_EXTERNAL_STORAGE: Int = 100
     }
 }
