@@ -2,16 +2,13 @@ package com.bassamworks.tunify.fragments
 
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.view.ViewPager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
 import com.bassamworks.tunify.R
 import com.bassamworks.tunify.adapters.pagerAdapters.LibraryPagerAdapter
 
-class LibraryFragment : Fragment() {
+class LibraryFragment : androidx.fragment.app.Fragment() {
 
     companion object {
         const val KEY_FRAGMENT_POSITION = "fragment_position"
@@ -24,7 +21,7 @@ class LibraryFragment : Fragment() {
 
         val view = inflater.inflate(R.layout.fragment_library, container, false)
 
-        view.findViewById<ViewPager>(R.id.view_pager_library).apply {
+        view.findViewById<androidx.viewpager.widget.ViewPager>(R.id.view_pager_library).apply {
             adapter = LibraryPagerAdapter(context, childFragmentManager)
             fragmentPosition?.let { setCurrentItem(it, true) }
         }
