@@ -2,14 +2,12 @@ package com.bassamworks.tunify.fragments
 
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
 import com.bassamworks.tunify.R
 
-class PlayQueueFragment : Fragment() {
+class PlayQueueFragment : androidx.fragment.app.Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
