@@ -1,40 +1,42 @@
 package com.bassamworks.tunify.fragments.libraryFragments
 
 
-import android.content.Context
-import android.database.Cursor
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
 import com.bassamworks.tunify.R
-import com.bassamworks.tunify.adapters.cursorAdapters.AlbumsCursorAdapter
-import com.bassamworks.tunify.excutors.AppExecutors
-import com.bassamworks.tunify.media.scanners.MediaStoreScanner
+import com.bassamworks.tunify.adapters.listAdapters.AlbumsListAdapter
+import com.bassamworks.tunify.models.ui.Album
+import com.bassamworks.tunify.viewModels.library.AlbumsViewModel
 import kotlinx.android.synthetic.main.fragment_albums.*
 
-class AlbumsFragment : Fragment() {
+class AlbumsFragment : LibraryFragment() {
+
+    private lateinit var model: AlbumsViewModel
+    private val rvAdapter = AlbumsListAdapter()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_albums, container, false)
+        val view = inflater.inflate(R.layout.fragment_albums, container, false)
+
+        model = ViewModelProviders.of(this)[AlbumsViewModel::class.java]
+
+        model.getLiveData().observe(this, Observer<List<Album>> {
+            updateAdapter(it ?: listOf())
+        })
+
+        return view
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        getAllAlbums(view.context)
+        rv_all_albums.adapter = rvAdapter
     }
 
-    private fun getAllAlbums(context: Context) {
-        AppExecutors.instance?.diskIO()?.execute {
-            updateAdapter(context, MediaStoreScanner(context).getAllAlbums())
-        }
+    private fun updateAdapter(albums: List<Album>) {
+        rvAdapter.submitList(albums)
     }
-
-    private fun updateAdapter(context: Context, cursor: Cursor) {
-        gv_all_albums.adapter = AlbumsCursorAdapter(context, cursor)
-    }
-
 }
