@@ -2,9 +2,8 @@ package com.bassamworks.tunify.media.scanners
 
 import android.content.Context
 import android.database.Cursor
-import android.provider.MediaStore
 import android.provider.MediaStore.Audio.*
-import android.provider.MediaStore.MediaColumns
+import com.bassamworks.tunify.constants.MediaStoreConstants
 import com.bassamworks.tunify.models.mediaStore.*
 import java.security.InvalidParameterException
 
@@ -12,140 +11,297 @@ class MediaStoreScanner(private val context: Context) {
 
     fun getAllSongs(): Cursor {
 
-        val contentUri = MediaStore.Audio.Media.EXTERNAL_CONTENT_URI
+        val contentUri = Media.EXTERNAL_CONTENT_URI
 
-        val selection = AudioColumns.IS_MUSIC + "!=?"
+        val selection = Media.IS_MUSIC + "!=?"
 
         val selectionArgs = arrayOf("0")
 
-        return context.contentResolver.query(contentUri, projectionSongs, selection, selectionArgs, null)
+        return context.contentResolver.query(contentUri, null, selection, selectionArgs, null)
     }
 
     fun getAllAlbums(): Cursor {
-        val contentUri = MediaStore.Audio.Albums.EXTERNAL_CONTENT_URI
+        val contentUri = Albums.EXTERNAL_CONTENT_URI
 
-        return context.contentResolver.query(contentUri, projectionAlbums, null, null, null)
+        return context.contentResolver.query(contentUri, null, null, null, null)
     }
 
     fun getAllArtists(): Cursor {
-        val contentUri = MediaStore.Audio.Artists.EXTERNAL_CONTENT_URI
+        val contentUri = Artists.EXTERNAL_CONTENT_URI
 
-        return context.contentResolver.query(contentUri, projectionArtists, null, null, null)
+        return context.contentResolver.query(contentUri, null, null, null, null)
     }
 
     fun getAllGenres(): Cursor {
-        val contentUri = MediaStore.Audio.Genres.EXTERNAL_CONTENT_URI
+        val contentUri = Genres.EXTERNAL_CONTENT_URI
 
-        return context.contentResolver.query(contentUri, projectionGenres, null, null, null)
+        return context.contentResolver.query(contentUri, null, null, null, null)
     }
 
     fun getAllPlaylists(): Cursor {
-        val contentUri = MediaStore.Audio.Playlists.EXTERNAL_CONTENT_URI
+        val contentUri = Playlists.EXTERNAL_CONTENT_URI
 
-        return context.contentResolver.query(contentUri, projectionPlaylists, null, null, null)
+        return context.contentResolver.query(contentUri, null, null, null, null)
     }
 
     /*fun getAllFolders(): Cursor {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }*/
 
-    private val projectionSongs = arrayOf(MediaColumns.DATA,
-            MediaColumns.SIZE,
-            MediaColumns.DISPLAY_NAME,
-            MediaColumns.TITLE,
-            MediaColumns.DATE_ADDED,
-            AudioColumns.ALBUM,
-            AudioColumns.ALBUM_ID,
-            AudioColumns.ARTIST,
-            AudioColumns.ARTIST_ID,
-            AudioColumns.BOOKMARK,
-            AudioColumns.COMPOSER,
-            AudioColumns.DURATION,
-            AudioColumns.TRACK,
-            AudioColumns.YEAR)
+    companion object {
 
-    private val projectionAlbums = arrayOf(AlbumColumns.ALBUM,
-            AlbumColumns.ALBUM_ART,
-            AlbumColumns.ALBUM_ID,
-            AlbumColumns.ARTIST,
-            AlbumColumns.FIRST_YEAR,
-            AlbumColumns.LAST_YEAR,
-            AlbumColumns.NUMBER_OF_SONGS)
+        fun getSongsFromCursor(cursor: Cursor): List<MSSong> {
+            if (cursor.isClosed || cursor.count <= 0) return listOf()
 
-    private val projectionArtists = arrayOf(ArtistColumns.ARTIST,
-            ArtistColumns.NUMBER_OF_ALBUMS,
-            ArtistColumns.NUMBER_OF_TRACKS)
+            val mediaStoreList: ArrayList<MSSong> = ArrayList(cursor.count)
 
-    private val projectionGenres = arrayOf(GenresColumns.NAME)
+            cursor.moveToPosition(-1)
 
-    private val projectionPlaylists = arrayOf(PlaylistsColumns.NAME,
-            PlaylistsColumns.DATA,
-            PlaylistsColumns.DATE_ADDED,
-            PlaylistsColumns.DATE_MODIFIED)
-}
+            while (cursor.moveToNext()) {
+                mediaStoreList.add(getSongFromCursor(cursor))
+            }
 
-fun getSongFromCursor(cursor: Cursor): MSSong {
-    if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+            return mediaStoreList
+        }
 
-    val filePath: String = cursor.getString(cursor.getColumnIndex(MediaColumns.DATA))
-    val size: Long = cursor.getLong(cursor.getColumnIndex(MediaColumns.SIZE))
-    val displayName: String = cursor.getString(cursor.getColumnIndex(MediaColumns.DISPLAY_NAME))
-    val title: String = cursor.getString(cursor.getColumnIndex(MediaColumns.TITLE))
-    val dateAdded: Long = cursor.getLong(cursor.getColumnIndex(MediaColumns.DATE_ADDED))
+        fun getAlbumsFromCursor(cursor: Cursor): List<MSAlbum> {
+            if (cursor.isClosed || cursor.count <= 0) return listOf()
 
-    val duration: Long = cursor.getLong(cursor.getColumnIndex(AudioColumns.DURATION))
-    val album: String = cursor.getString(cursor.getColumnIndex(AudioColumns.ALBUM))
-    val albumId: Long = cursor.getLong(cursor.getColumnIndex(AudioColumns.ALBUM_ID))
-    val artist: String = cursor.getString(cursor.getColumnIndex(AudioColumns.ARTIST))
-    val artistId: Long = cursor.getLong(cursor.getColumnIndex(AudioColumns.ARTIST_ID))
-    val lastPlaybackBookmark: Long = cursor.getLong(cursor.getColumnIndex(AudioColumns.BOOKMARK))
-    val composer: String = cursor.getString(cursor.getColumnIndex(AudioColumns.COMPOSER))
-    val trackNumber: Int = cursor.getInt(cursor.getColumnIndex(AudioColumns.TRACK))
-    val year: Int = cursor.getInt(cursor.getColumnIndex(AudioColumns.YEAR))
+            val mediaStoreList: ArrayList<MSAlbum> = ArrayList(cursor.count)
 
-    return MSSong(filePath, size, displayName, title, dateAdded, duration, album, albumId,
-            artist, artistId, lastPlaybackBookmark, composer, trackNumber, year)
-}
+            cursor.moveToPosition(-1)
 
-fun getAlbumFromCursor(cursor: Cursor): MSAlbum {
-    if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+            while (cursor.moveToNext()) {
+                mediaStoreList.add(getAlbumFromCursor(cursor))
+            }
 
-    val album: String = cursor.getString(cursor.getColumnIndex(AlbumColumns.ALBUM))
-    val albumArt: String = cursor.getString(cursor.getColumnIndex(AlbumColumns.ALBUM_ART))
-    val albumId: Long = cursor.getLong(cursor.getColumnIndex(AlbumColumns.ALBUM_ID))
-    val artist: String = cursor.getString(cursor.getColumnIndex(AlbumColumns.ARTIST))
-    val firstYear: Int = cursor.getInt(cursor.getColumnIndex(AlbumColumns.FIRST_YEAR))
-    val lastYear: Int = cursor.getInt(cursor.getColumnIndex(AlbumColumns.LAST_YEAR))
-    val numberOfSongs: Int = cursor.getInt(cursor.getColumnIndex(AlbumColumns.NUMBER_OF_SONGS))
+            return mediaStoreList
+        }
 
-    return MSAlbum(album, albumArt, albumId, artist, firstYear, lastYear, numberOfSongs)
-}
+        fun getArtistsFromCursor(cursor: Cursor): List<MSArtist> {
+            if (cursor.isClosed || cursor.count <= 0) return listOf()
 
-fun getArtistFromCursor(cursor: Cursor): MSArtist {
-    if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+            val mediaStoreList: ArrayList<MSArtist> = ArrayList(cursor.count)
 
-    val artist: String = cursor.getString(cursor.getColumnIndex(ArtistColumns.ARTIST))
-    val numberOfAlbums: Int = cursor.getInt(cursor.getColumnIndex(ArtistColumns.NUMBER_OF_ALBUMS))
-    val numberOfTracks: Int = cursor.getInt(cursor.getColumnIndex(ArtistColumns.NUMBER_OF_TRACKS))
+            cursor.moveToPosition(-1)
 
-    return MSArtist(artist, numberOfAlbums, numberOfTracks)
-}
+            while (cursor.moveToNext()) {
+                mediaStoreList.add(getArtistFromCursor(cursor))
+            }
 
-fun getGenreFromCursor(cursor: Cursor): MSGenre {
-    if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+            return mediaStoreList
+        }
 
-    val genre: String = cursor.getString(cursor.getColumnIndex(GenresColumns.NAME))
+        fun getGenresFromCursor(cursor: Cursor): List<MSGenre> {
+            if (cursor.isClosed || cursor.count <= 0) return listOf()
 
-    return MSGenre(genre)
-}
+            val mediaStoreList: ArrayList<MSGenre> = ArrayList(cursor.count)
 
-fun getPlaylistFromCursor(cursor: Cursor): MSPlaylist {
-    if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+            cursor.moveToPosition(-1)
 
-    val name: String = cursor.getString(cursor.getColumnIndex(PlaylistsColumns.NAME))
-    val playlistFilePath: String = cursor.getString(cursor.getColumnIndex(PlaylistsColumns.DATA))
-    val playlistDateAdded: Long = cursor.getLong(cursor.getColumnIndex(PlaylistsColumns.DATE_ADDED))
-    val playlistDateModified: Long = cursor.getLong(cursor.getColumnIndex(PlaylistsColumns.DATE_MODIFIED))
+            while (cursor.moveToNext()) {
+                mediaStoreList.add(getGenreFromCursor(cursor))
+            }
 
-    return MSPlaylist(name, playlistFilePath, playlistDateAdded, playlistDateModified)
+            return mediaStoreList
+        }
+
+        fun getPlaylistsFromCursor(cursor: Cursor): List<MSPlaylist> {
+            if (cursor.isClosed || cursor.count <= 0) return listOf()
+
+            val mediaStoreList: ArrayList<MSPlaylist> = ArrayList(cursor.count)
+
+            cursor.moveToPosition(-1)
+
+            while (cursor.moveToNext()) {
+                mediaStoreList.add(getPlaylistFromCursor(cursor))
+            }
+
+            return mediaStoreList
+        }
+
+        fun getSongFromCursor(cursor: Cursor): MSSong {
+            if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+
+            val id: Long = cursor.getLong(cursor.getColumnIndex(Media._ID))
+
+            val filePath: String = cursor.getString(cursor.getColumnIndex(Media.DATA))
+            val size: Long = cursor.getLong(cursor.getColumnIndex(Media.SIZE))
+            val displayName: String = cursor.getString(cursor.getColumnIndex(Media.DISPLAY_NAME))
+            val title: String = cursor.getString(cursor.getColumnIndex(Media.TITLE))
+
+            val durationIndex = cursor.getColumnIndex(Media.DURATION)
+            val duration: Long? = if (durationIndex != INVALID_COLUMN) cursor.getLong(durationIndex) else null
+
+            val albumIndex = cursor.getColumnIndex(Media.ALBUM)
+            val album: String? = if (albumIndex != INVALID_COLUMN) cursor.getString(albumIndex) else null
+
+            val albumIdIndex = cursor.getColumnIndex(Media.ALBUM_ID)
+            val albumId: Long? = if (albumIdIndex != INVALID_COLUMN) cursor.getLong(albumIdIndex) else null
+
+            val artistIndex = cursor.getColumnIndex(Media.ARTIST)
+            val artist: String? = if (artistIndex != INVALID_COLUMN) cursor.getString(artistIndex) else null
+
+            val artistIdIndex = cursor.getColumnIndex(Media.ARTIST_ID)
+            val artistId: Long? = if (artistIdIndex != INVALID_COLUMN) cursor.getLong(artistIdIndex) else null
+
+            val lastPlaybackBookmarkIndex = cursor.getColumnIndex(Media.BOOKMARK)
+            val lastPlaybackBookmark: Long? = if (lastPlaybackBookmarkIndex != INVALID_COLUMN) cursor.getLong(lastPlaybackBookmarkIndex) else null
+
+            val composerIndex = cursor.getColumnIndex(Media.COMPOSER)
+            val composer: String? = if (composerIndex != INVALID_COLUMN) cursor.getString(composerIndex) else null
+
+            val trackNumberIndex = cursor.getColumnIndex(Media.TRACK)
+            val trackNumber: Int? = if (trackNumberIndex != INVALID_COLUMN) cursor.getInt(trackNumberIndex) else null
+
+            val yearIndex = cursor.getColumnIndex(Media.YEAR)
+            val year: Int? = if (yearIndex != INVALID_COLUMN) cursor.getInt(yearIndex) else null
+
+            return MSSong(id, filePath, size, displayName, title,
+                    duration ?: MediaStoreConstants.DEFAULT_DURATION,
+                    album ?: MediaStoreConstants.DEFAULT_ALBUM_NAME,
+                    albumId ?: MediaStoreConstants.DEFAULT_ALBUM_ID,
+                    artist ?: MediaStoreConstants.DEFAULT_ARTIST_NAME,
+                    artistId ?: MediaStoreConstants.DEFAULT_ARTIST_ID,
+                    lastPlaybackBookmark ?: MediaStoreConstants.DEFAULT_PLAYBACK_BOOKMARK,
+                    composer ?: MediaStoreConstants.DEFAULT_COMPOSER,
+                    trackNumber ?: MediaStoreConstants.DEFAULT_TRACK_NUMBER,
+                    year ?: MediaStoreConstants.DEFAULT_YEAR)
+        }
+
+
+        fun getAlbumFromCursor(cursor: Cursor): MSAlbum {
+            if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+
+            val id: Long = cursor.getLong(cursor.getColumnIndex(Albums._ID))
+
+            val albumIndex = cursor.getColumnIndex(Albums.ALBUM)
+            val album: String? = if (albumIndex != INVALID_COLUMN) cursor.getString(albumIndex) else null
+
+            val albumArtIndex = cursor.getColumnIndex(Albums.ALBUM_ART)
+            val albumArt: String? = if (albumArtIndex != INVALID_COLUMN) cursor.getString(albumArtIndex) else null
+
+            val albumIdIndex = cursor.getColumnIndex(Albums.ALBUM_ID)
+            val albumId: Long? = if (albumIdIndex != INVALID_COLUMN) cursor.getLong(albumIdIndex) else null
+
+            val artistIndex = cursor.getColumnIndex(Albums.ARTIST)
+            val artist: String? = if (artistIndex != INVALID_COLUMN) cursor.getString(artistIndex) else null
+
+            val firstYearIndex = cursor.getColumnIndex(Albums.FIRST_YEAR)
+            val firstYear: Int? = if (firstYearIndex != INVALID_COLUMN) cursor.getInt(firstYearIndex) else null
+
+            val lastYearIndex = cursor.getColumnIndex(Albums.LAST_YEAR)
+            val lastYear: Int? = if (lastYearIndex != INVALID_COLUMN) cursor.getInt(lastYearIndex) else null
+
+            val numberOfSongsIndex = cursor.getColumnIndex(Albums.NUMBER_OF_SONGS)
+            val numberOfSongs: Int? = if (numberOfSongsIndex != INVALID_COLUMN) cursor.getInt(numberOfSongsIndex) else null
+
+            return MSAlbum(id,
+                    album ?: MediaStoreConstants.DEFAULT_ALBUM_NAME,
+                    albumArt ?: MediaStoreConstants.DEFAULT_ALBUM_ART,
+                    albumId ?: MediaStoreConstants.DEFAULT_ALBUM_ID,
+                    artist ?: MediaStoreConstants.DEFAULT_ARTIST_NAME,
+                    firstYear ?: MediaStoreConstants.DEFAULT_FIRST_YEAR,
+                    lastYear ?: MediaStoreConstants.DEFAULT_LAST_YEAR,
+                    numberOfSongs ?: MediaStoreConstants.DEFAULT_TRACK_NUMBER)
+        }
+
+
+        fun getArtistFromCursor(cursor: Cursor): MSArtist {
+            if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+
+            val id: Long = cursor.getLong(cursor.getColumnIndex(Artists._ID))
+
+            val artistIndex = cursor.getColumnIndex(Artists.ARTIST)
+            val artist: String? = if (artistIndex != INVALID_COLUMN) cursor.getString(artistIndex) else null
+
+            val numberOfAlbumsIndex = cursor.getColumnIndex(Artists.NUMBER_OF_ALBUMS)
+            val numberOfAlbums: Int? = if (numberOfAlbumsIndex != INVALID_COLUMN) cursor.getInt(numberOfAlbumsIndex) else null
+
+            val numberOfTracksIndex = cursor.getColumnIndex(Artists.NUMBER_OF_TRACKS)
+            val numberOfTracks: Int? = if (numberOfTracksIndex != INVALID_COLUMN) cursor.getInt(numberOfTracksIndex) else null
+
+            return MSArtist(id,
+                    artist ?: MediaStoreConstants.DEFAULT_ARTIST_NAME,
+                    numberOfAlbums ?: MediaStoreConstants.DEFAULT_NUMBER_OF_ALBUMS,
+                    numberOfTracks ?: MediaStoreConstants.DEFAULT_TRACK_NUMBER)
+        }
+
+
+        fun getGenreFromCursor(cursor: Cursor): MSGenre {
+            if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+
+            val id: Long = cursor.getLong(cursor.getColumnIndex(Genres._ID))
+
+            val genreIndex = cursor.getColumnIndex(Genres.NAME)
+            val genre: String? = if (genreIndex != INVALID_COLUMN) cursor.getString(genreIndex) else null
+
+            return MSGenre(id,
+                    genre ?: MediaStoreConstants.DEFAULT_GENRE)
+        }
+
+
+        fun getPlaylistFromCursor(cursor: Cursor): MSPlaylist {
+            if (cursor.isClosed || cursor.count == 0) throw InvalidParameterException("Invalid cursor")
+
+            val id: Long = cursor.getLong(cursor.getColumnIndex(Playlists._ID))
+
+            val nameIndex = cursor.getColumnIndex(Playlists.NAME)
+            val name: String? = if (nameIndex != INVALID_COLUMN) cursor.getString(nameIndex) else null
+
+            val playlistFilePathIndex = cursor.getColumnIndex(Playlists.DATA)
+            val playlistFilePath: String? = if (playlistFilePathIndex != INVALID_COLUMN) cursor.getString(playlistFilePathIndex) else null
+
+            val playlistDateAddedIndex = cursor.getColumnIndex(Playlists.DATE_ADDED)
+            val playlistDateAdded: Long? = if (playlistDateAddedIndex != INVALID_COLUMN) cursor.getLong(playlistDateAddedIndex) else null
+
+            val playlistDateModifiedIndex = cursor.getColumnIndex(Playlists.DATE_MODIFIED)
+            val playlistDateModified: Long? = if (playlistDateModifiedIndex != INVALID_COLUMN) cursor.getLong(playlistDateModifiedIndex) else null
+
+            return MSPlaylist(id,
+                    name ?: MediaStoreConstants.DEFAULT_ARTIST_NAME,
+                    playlistFilePath ?: MediaStoreConstants.DEFAULT_PLAYLIST_PATH,
+                    playlistDateAdded ?: MediaStoreConstants.DEFAULT_PLAYLIST_DATE_ADDED,
+                    playlistDateModified ?: MediaStoreConstants.DEFAULT_PLAYLIST_DATE_MODIFIED)
+        }
+
+        private const val INVALID_COLUMN = -1
+    }
+
+    private val projectionSongs = arrayOf(Media._ID,
+            Media.DATA,
+            Media.SIZE,
+            Media.DISPLAY_NAME,
+            Media.TITLE,
+            Media.ALBUM,
+            Media.ALBUM_ID,
+            Media.ARTIST,
+            Media.ARTIST_ID,
+            Media.BOOKMARK,
+            Media.COMPOSER,
+            Media.DURATION,
+            Media.TRACK,
+            Media.YEAR)
+
+    private val projectionAlbums = arrayOf(Albums._ID,
+            Albums.ALBUM,
+            Albums.ALBUM_ART,
+            Albums.ALBUM_ID,
+            Albums.ARTIST,
+            Albums.FIRST_YEAR,
+            Albums.LAST_YEAR,
+            Albums.NUMBER_OF_SONGS)
+
+    private val projectionArtists = arrayOf(Artists._ID,
+            Artists.ARTIST,
+            Artists.NUMBER_OF_ALBUMS,
+            Artists.NUMBER_OF_TRACKS)
+
+    private val projectionGenres = arrayOf(Genres._ID, Genres.NAME)
+
+    private val projectionPlaylists = arrayOf(Playlists._ID,
+            Playlists.NAME,
+            Playlists.DATA,
+            Playlists.DATE_ADDED,
+            Playlists.DATE_MODIFIED)
 }
