@@ -1,7 +1,9 @@
 package com.bassamworks.tunify.models.mediaStore
 
-open class MSAudioFile(open val filePath: String,
+import com.bassamworks.tunify.models.IModelType
+
+open class MSAudioFile(open val id: Long,
+                       open val filePath: String,
                        open val size: Long,
                        open val displayName: String,
-                       open val title: String,
-                       open val dateAdded: Long)
+                       open val title: String) : IModelType.MediaStore
