@@ -1,18 +1,26 @@
 package com.bassamworks.tunify.models.mediaStore
 
 import com.bassamworks.tunify.constants.MediaStoreConstants
+import com.bassamworks.tunify.models.IModelType
+import com.bassamworks.tunify.models.ui.Song
 
-data class MSSong(override val filePath: String,
+data class MSSong(override val id: Long,
+                  override val filePath: String,
                   override val size: Long,
                   override val displayName: String,
                   override val title: String,
-                  override val dateAdded: Long,
-                  val duration: Long,
+                  val duration: Long = MediaStoreConstants.DEFAULT_DURATION,
                   val album: String = MediaStoreConstants.DEFAULT_ALBUM_NAME,
                   val albumId: Long = MediaStoreConstants.DEFAULT_ALBUM_ID,
                   val artist: String = MediaStoreConstants.DEFAULT_ARTIST_NAME,
                   val artistId: Long = MediaStoreConstants.DEFAULT_ARTIST_ID,
-                  val lastPlaybackBookmark: Long = 0,
+                  val lastPlaybackBookmark: Long = MediaStoreConstants.DEFAULT_PLAYBACK_BOOKMARK,
                   val composer: String = MediaStoreConstants.DEFAULT_COMPOSER,
                   val trackNumber: Int = MediaStoreConstants.DEFAULT_TRACK_NUMBER,
-                  val year: Int = MediaStoreConstants.DEFAULT_YEAR) : MSAudioFile(filePath, size, displayName, title, dateAdded)
+                  val year: Int = MediaStoreConstants.DEFAULT_YEAR)
+    : MSAudioFile(id, filePath, size, displayName, title), IModelType.MediaStore {
+
+    fun toUIModel() = Song(id, filePath, title, album, artist, duration.toDouble(), lastPlaybackBookmark.toDouble())
+
+}
+
